Conflicts:
	bsd-user/main.c
	roms/seabios
	target-mips/machine.c
	ui/vnc.c
	user-exec.c

+++ hw/tpm/tpm_tis.h
@@ -41,7 +41,7 @@ typedef enum {
 typedef struct TPMLocality {
     TPMTISState state;
     uint8_t access;
-    uint8_t sts;
+    uint32_t sts;
     uint32_t inte;
     uint32_t ints;
